用户扩展

+++ my_blog/article/views.py
@@ -6,20 +6,32 @@ import article
 from .models import ArticlePost
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
 
 
 def article_list(request):
     # 取出所有博客文章
-    articles = ArticlePost.objects.all()
-    # 需要传递给模板（templates）的对象
+    article_list = ArticlePost.objects.all()
+
+    # 每页显示 1 篇文章
+    paginator = Paginator(article_list, 3)
+    # 获取 url 中的页码
+    page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
+
     context = {'articles': articles}
-    # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
 
 def article_detail(request, id):
     # 取出相应的文章
     article = ArticlePost.objects.get(id=id)
+
+    # 浏览量 +1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
 
     # 将markdown语法渲染成html样式
     article.body = markdown.markdown(article.body,
@@ -38,6 +50,7 @@ def article_detail(request, id):
 
 
 # 写文章的视图
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 判断用户是否提交数据
     if request.method == "POST":
@@ -47,7 +60,7 @@ def article_create(request):
             # 保存数据，但暂时不提交到数据库中
             new_article = article_post_form.save(commit=False)
             # 指定数据库中 id=1 的用户为作者
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存到数据库中
             new_article.save()
             # 完成后返回到文章列表
@@ -66,6 +79,7 @@ def article_create(request):
 
 
 # 删文章
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     article = ArticlePost.objects.get(id=id)
     article.delete()
@@ -73,9 +87,14 @@ def article_delete(request, id):
 
 
 # 安全删除文章
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
+    article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权删除这篇文章。")
+
     if request.method == 'POST':
-        article = ArticlePost.objects.get(id=id)
         article.delete()
         return redirect("article:article_list")
     else:
@@ -83,6 +102,7 @@ def article_safe_delete(request, id):
 
 
 # 更新文章
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -93,6 +113,11 @@ def article_update(request, id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
+
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
@@ -104,7 +129,7 @@ def article_update(request, id):
             article.body = request.POST['body']
             article.save()
             # 完成后返回到修改后的文章中。需传入文章的 id 值
-            return redirect('article:article_detail',id=id)
+            return redirect('article:article_detail', id=id)
         else:
             return HttpResponse("表单内容有误，请重新填写。")
     # 如果用户 GET 请求获取数据
@@ -112,6 +137,6 @@ def article_update(request, id):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文，将 article 文章对象也传递进去，以便提取旧的内容
-        context = { 'article': article, 'article_post_form': article_post_form }
+        context = {'article': article, 'article_post_form': article_post_form}
         # 将响应返回到模板中
         return render(request, 'article/update.html', context)
